+++ client/App.js
@@ -1,29 +1,19 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View, Button } from 'react-native';
 import Auth from './components/Auth';
-import React, {useState, useEffect} from 'react';
+import React, {useState, useCallback} from 'react';
+import { useFocusEffect } from '@react-navigation/native';
+import { NavigationContainer } from '@react-navigation/native';
 
 export default function App() {
-  const [statusBarToggle, setStatusBarToggle] = useState(true)
 
-  const handleButtonPress = () => {
-    // Action to perform when the button is pressed
-    console.log('Button pressed!');
-  };
-
-  const handleHide = () => {
-    StatusBar.setHidden(true, 'slide')
-    console.log('handlehide')
-  }
 
   return (
-    <View style={styles.container}>
-      <Auth />
-      <Text>This is the APP component.</Text>
-      <Text>Blah Blah Blah Blah Blah</Text>
-      <Button title="Random button!" onPress={handleButtonPress} />
-      <StatusBar style="auto" onPress={handleHide}/>
-    </View>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen name="Home" component={HomeScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
